BPE26-ref: extract progress bar feature to dedicated file
