+++ src/features/ui/index.ts
@@ -1,2 +1,4 @@
 export { createFullscreenButtonSection } from "./fullscreen-button";
 export { setupPlayerStickiness } from "./player-stickiness";
+export { createProgressBar, updateProgressBar } from "./progress";
+export { createVolumeControlSection, handleMuteToggle, syncMuteBtn } from "./volume";
